Move PR test to test_PR.py file (#444)
* move PR test from test_issues.py to test_PR.py
